Restructured project tree
This restructuring is required since we need to have node.go on the
same level with pbft-server.go, txlist.go and txpool.go. We need to
be able to access the transaction pool inside the pbft phases defined
in node.go. Trying to access the transaction pool through the existing
structure would lead to cyclic importing of packages.

Signed-off-by: Samikshan Bairagya <samikshan@gmail.com>